Add files via upload
This GitHub repository contains the source code of an office application designed to simplify the management of school activities. The app offers features such as student management, courses, timetables, and many other aspects of school life. Developed with passion, this application aims to improve efficiency and organization within the school.